requirements.txt
streamlit==1.32.0
google-generativeai==0.3.2
pdfplumber==0.10.3
pytesseract==0.3.10
plotly==5.18.0
altair==5.2.0
spacy==3.7.2
python-dotenv==1.0.1
Pillow==10.2.0
GEMINI_API_KEY=your_api_key_here 